Create new views: add_record, update_record

+++ website/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
-from .forms import SignUpForm
+from .forms import SignUpForm, AddRecordForm
 from .models import Record
 
 
@@ -69,7 +69,37 @@ def customer_record(request, pk):
         return redirect('home')
     
 def delete_record(request, pk):
-    delete_it = Record.objects.get(id = pk)
-    delete_it.delete()
-    messages.success(request, "Klient usunięty poprawnie.")
-    return redirect('home')
+    if request.user.is_authenticated:
+        delete_it = Record.objects.get(id = pk)
+        delete_it.delete()
+        messages.success(request, "Klient usunięty poprawnie.")
+        return redirect('home')
+    else:
+        messages.success(request, "Aby zobaczyć tę stronę, musisz być zalogowany")
+        return redirect('home')
+    
+def add_record(request):
+    form = AddRecordForm(request.POST or None)
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            if form.is_valid():
+                add_record = form.save()
+                messages.success(request, "Dodano klienta poprawnie.")
+                return redirect('home')
+        return render(request, 'add_record.html', {'form': form})
+    else:
+        messages.success(request, "Musisz być zalogowany. Zaloguj się i spróbuj ponownie.")
+        return redirect('home')
+    
+def update_record(request, pk):
+    if request.user.is_authenticated:
+        update_it = Record.objects.get(id = pk)
+        form = AddRecordForm(request.POST or None, instance=update_it)
+        if form.is_valid():
+            form.save()
+            messages.success(request, "Zaktualizowano dane klienta poprawnie.")
+            return redirect('home')
+        return render(request, 'update_record.html', {'form': form})
+    else:
+        messages.success(request, "Musisz być zalogowany. Zaloguj się i spróbuj ponownie.")
+        return redirect('home')
